・アクションゲームに移行

+++ nkEngine/_Math/nkHash.h
@@ -8,7 +8,7 @@ namespace nkEngine
 		static int MakeHash(const char* string)
 		{
 			int hash = 0;
-			int len = strlen(string);
+			int len = (int)strlen(string);
 			for (int i = 0; i < len; i++)
 			{
 				hash = hash * 37 + string[i];
